Renamed `muller_genotypes` to `clusters`
