+++ muller/muller_output/save_output.py
@@ -12,8 +12,8 @@ FILTERED_GENOTYPE_LABEL = "removed"
 OutputType = Tuple[pandas.DataFrame, pandas.DataFrame, str, Dict[str, Any]]
 
 try:
-	from muller.muller_genotypes.generate import GenotypeOptions
-	from muller_genotypes.metrics.pairwise_calculation_cache import PairwiseCalculation
+	from muller.clustering.generate import GenotypeOptions
+	from clustering.metrics.pairwise_calculation_cache import PairwiseCalculation
 	from inheritance.sort_genotypes import SortOptions
 	from inheritance.order import OrderClusterParameters
 	from graphics import plot_genotypes, plot_heatmap, plot_dendrogram, generate_muller_plot
@@ -21,7 +21,7 @@ try:
 	from muller.muller_output.generate_scripts import generate_mermaid_diagram, generate_r_script, excecute_mermaid_script, execute_r_script
 	from muller.widgets import generate_genotype_palette, map_trajectories_to_genotype
 except ModuleNotFoundError:
-	from muller_genotypes.metrics.pairwise_calculation_cache import PairwiseCalculation
+	from clustering.metrics.pairwise_calculation_cache import PairwiseCalculation
 	from graphics import plot_genotypes, plot_heatmap, plot_dendrogram, generate_muller_plot
 	from muller_output.generate_tables import *
 	from muller_output.generate_scripts import generate_mermaid_diagram, generate_r_script, excecute_mermaid_script, execute_r_script
