+++ tests/test_genotype_average.py
@@ -1,6 +1,6 @@
 import pytest
 from import_data import import_table_from_string
-from muller_genotypes.average import _calculate_mean_frequencies_of_trajectories, calculate_mean_genotype
+from clustering.average import _calculate_mean_frequencies_of_trajectories, calculate_mean_genotype
 import pandas
 
 @pytest.fixture
